Adapting Go netcode example to Bootstrapping
